move array transforms to a separate file
